Switched all position size and order sizes to int. Made OrderAlgorithm use longs for mathing == and != for accuracy of synchronization. Also, added additional PerformCompare() calls for OrderAlgorithm when filled, canceled, etc. events occur.

+++ Platform/TickZoomAPI1.0/Engine/LogicalOrder.cs
@@ -51,7 +51,7 @@ namespace TickZoom.Api
 			set;
 		}
 		
-		double StrategyPosition {
+		int StrategyPosition {
 			get;
 			set;
 		}
@@ -71,7 +71,7 @@ namespace TickZoom.Api
 			set;
 		}
 		
-		double Positions {
+		int Positions {
 			get;
 			set;
 		}
