+++ Platform/TickZoomAPI1.0/Provider/PhysicalOrder.cs
@@ -56,7 +56,7 @@ namespace TickZoom.Api
 			get;
 		}
 		
-		double Size {
+		int Size {
 			get;
 			set;
 		}
